GEODE-3539: Additional test refactoring
Separated remaining tests from ConfigCommandDUnitTest into command specific classes.
